Intent Recognition Protoype/requirements.txt
absl-py==0.9.0
alembic==1.4.2
astunparse==1.6.3
audioread==2.1.8
boto3==1.13.26
botocore==1.16.26
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cliff==3.2.0
cmaes==0.5.0
cmd2==0.8.9
colorlog==4.1.0
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
deprecation==2.1.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dominate==2.5.1
eventlet==0.25.2
eyeD3==0.9.5
filelock==3.0.12
filetype==1.0.7
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.3
future==0.18.2
gast==0.3.3
gevent==20.5.2
gevent-websocket==0.10.1
google-api-core==1.17.0
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.29.0
h5py==2.10.0
hmmlearn==0.2.3
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.5.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.33.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.2
monotonic==1.5
mpmath==1.1.0
numba==0.50.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
optuna==1.5.0
packaging==20.4
pandas==1.0.4
pbr==5.4.5
pescador==2.1.0
Pillow==7.1.2
plotly==4.8.1
prettytable==0.7.2
protobuf==3.12.2
pyannote.audio==2.0a1
pyannote.core==4.1
pyannote.database==3.0.1
pyannote.metrics==2.3
pyannote.pipeline==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.0
python-socketio==4.6.0
pytorch-pretrained-bert==0.6.2
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
sentencepiece==0.1.91
simplejson==3.16.0
six==1.15.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.17
stevedore==2.0.0
sympy==1.6
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.7.0
torch==1.5.0
tqdm==4.46.1
transformers==2.11.0
typing-extensions==3.7.4.2
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.2.4
websocket==0.2.1
websocket-client==0.48.0
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
